refactor: move analytics to its own client
