+++ src/modules/stats/sagas.ts
@@ -1,5 +1,6 @@
 import { BuilderAPI } from 'lib/api/builder'
 import { call, put, takeEvery } from 'redux-saga/effects'
+import { analytics } from 'lib/api/analytics'
 import {
   fetchWeeklySceneStatsFailure,
   FetchWeeklySceneStatsRequestAction,
@@ -8,13 +9,13 @@ import {
 } from './action'
 import { WeeklyStats } from './types'
 
-export function* statsSaga(builder: BuilderAPI) {
+export function* statsSaga(_builder: BuilderAPI) {
   yield takeEvery(FETCH_WEEKLY_SCENE_STATS_REQUEST, handleFetchSceneStatsRequest)
 
   function* handleFetchSceneStatsRequest(action: FetchWeeklySceneStatsRequestAction) {
     const { base } = action.payload
     try {
-      const stats: WeeklyStats = yield call(() => builder.fetchWeeklyStats(base))
+      const stats: WeeklyStats = yield call(() => analytics.fetchWeeklyStats(base))
       yield put(fetchWeeklySceneStatsSuccess(base, stats))
     } catch (error) {
       yield put(fetchWeeklySceneStatsFailure(base, error.message))
